+++ NEWS.md
@@ -2,12 +2,12 @@
 
 ## Major changes
 
-- Factor valued predictions are now supported. Each level is represented by its dummy variable.
+- Factor-valued predictions are now supported. Each level is represented by its dummy variable.
 
 ## Other changes
 
 - Slight speed-up.
-- Integer valued case-weights are now turned into doubles to avoid integer overflow.
+- Integer valued case weights are now turned into doubles to avoid integer overflow.
 
 # kernelshap 0.3.8